fix: Renames section to collection
While trying to keep the scope of this PR as small as possible, I went through and renamed gateway endpoints to use `collection` instead of `section` to better match the changes to `defineSchema()`.

A non-exhaustive list includes:
* datasource
  * getTemplatesForCollection
  * getDocumentsForCollection
  * getSettingsForCollection
  * getCollectionsSettings
  * getCollection
  * getCollectionByPath
* resolver
  * getCollection & resolveCollection
  * getCollections & resolveCollections
* sys (systemInfo)
  * collection
* addPendingDocumentMutation

I also updated the `useGraphQLForm` and `useDocumentCreatorPlugin` to also refer to things as `collection` rather than `section`.

Closes #167