feat: Add filtering to collections in Add Document
Jeff had the idea that it would be better Dx if we allowed for filtering of `collections` in the "Add Document" content creator.

This just adds an optional param (`filterCollections`) to `useDocumentCreatorPlugin` that provides the list of `options` and returns a (potentially) filtered list of `options`.

I also updated the CONTRIBUTING readme after we learned that yarn v2+ is not supported with Node >14.