273 - Revisits useDocumentCreatorPlugin to improve UX
The initial version of `useDocumentCreatorPlugin()` cheated a bit with Tina Standards, leveraging the constant calling of `validate()` to make it behave like an `onChange()` event with controlled fields.  This meant adding a bunch of _non-validation_ logic into the functions for `collection` and `collectionTemplate` as well as embedding the value for `collection` inside `collectionTemplate`.

Ideally, we'd want `collection` and `template` to be deterministic, but separate.

This change leverages the recent changes to `ContentCreator` Plugin that lets us use `onChange()` and `initialValues`, removing much of the `validate()` logic and utilizing `useState()` to watch for changes to `collection` and determining the `options` for `template` from that.

It is not completely void of "questionable" code - particularly saving the `plugin` in `useState()` as well and having a separate `useEffect()` for `.add()` and `.remove()`.

But it cleans up the UX for the `useDocumentCreatorPlugin()` and, hopefully, more closely adheres to the Tina Way (tm).